Make it impossible to open the progress view sheet multiple times
Summary:
It used to be possible after resizing the window so narrowly that the drawer switched into modal mode and back to its original width.
There is still the question of why the sheet is sometimes left behind the drawer, but that's probably something for another patch.

Test Plan: Multiple ProgressView sheets can't be opened at the same time anymore

Reviewers: #discover_software_store, apol

Reviewed By: #discover_software_store, apol

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D25109