Add support for sub-comments to the Reviews backend
Summary:
Using a similar style to the comments delegate in KNewStuffQuick, this
adds the functionality for having subcomments in the Reviews model.
The end result is that we no longer have incorrect numbers showing in
Discover, as it now shows all the items returned by the server.

* Adds a role for the depth of reviews (used for comments on reviews)
* Adds sub-comments (and set their depth) in the KNS backend's reviews
* Adapts the ReviewDelegate to have indentation marks (styled like KNSQuick)

BUG: 391191
FIXED-IN: 5.17.0

Test Plan:
Without the patch, incorrect numbers of reviews are shown (as no children are shown). With this patch, children are shown (using the same indentation style as used by KNewStuffQuick):
{F7812653}

Reviewers: #discover_software_store, #plasma, apol, ngraham

Reviewed By: #discover_software_store, ngraham

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D25886