Improve layout and appearance of app details view
Summary:
After porting to FormLayout in D27183, the layout can now be so wide due to the URLs
that it switches to the compact/mobile view. This patch re-arranges things a bit and
uses human-readable strings for the links instead of showing raw URLs. For safety,
this requires D27663 to be implemented first.

BUG: 418019
FIXED-IN: 5.19.0
Depends on D27663

Test Plan:
Before: normal-sized window (shows the compact/mobile view): {F8131970}

Before wide window: {F8131971}

After: normal-sized or wide window: {F8131972}

Reviewers: #vdg, #discover_software_store, apol, ndavis

Reviewed By: #vdg, #discover_software_store, apol, ndavis

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D27664