packagekit: Fix checking of APT::Periodic::Update-Package-Lists
Summary:
BUG: 419055

Currently `PackageKitNotifier` attempts to check the frequency by running `apt-config dump` and looking for the `Apt::Periodic::Update-Package-Lists` property in its output. However, on my system the property name starts with uppercase `APT` rather than `Apt`, and ends with a semicolon.

This patch is a defensive, backward-compatible fix. It assumes that under some circumstances the previous code was correct to look for `Apt` and no ending semicolon. To cover both cases, I edited the regexp to allow for an optional semicolon, and to make the comparison case-insensitive.

Reviewers: #discover_software_store, apol, ngraham

Reviewed By: #discover_software_store, apol

Subscribers: plasma-devel

Tags: #plasma, #discover_software_store

Differential Revision: https://phabricator.kde.org/D28167