[Notifier] Port from KRun::runCommand
Use ApplicationLauncherJob for launching the main application and CommandLauncherJob for launching the updater.
This also gives us a nice startup feedback for both cases now.
The KService::serviceByDesktopName could in theory return null but imho that would be severely broken
if we have the updater running but rest of Discover is absent.

Differential Revision: https://phabricator.kde.org/D28821