Load the KNewStuff updates information on Discover startup
Without this, Discover only loads the installed items, but does not
also check whether there are updates available. This is cached by
the KNSCore::Engine, so we don't need the throttling in Discover itself,
but also if this is not done during the initialisation step, the updates
information will simply not be available.

BUG:421924
BUG:417987