Make basic KNS updating (and installing multi-item entries) work
Prior to this, the only KNS entries which would be able to install
correctly were entries with exactly one download item in them. This
patch fixes this part of the issue.

The second part of a fix, to ensure things work correctly, is to forward
questions from KNS correctly, but this will require deeper intervention
in Discover, to allow proceed requests to also offer options other than
a yes/no style thing (specifically for this purpose, we need to be able
to pick from a list of items).