Fix a crash in flatpak update notifier
Currently, the update notifier will crash due to a segmentation fault if
a flatpak package has been updated.

The segmentation fault occurs because the installationChanged() callback
expects the user data to have a type of FlatpakNotifier, but in reality
it's FlatpakNotifier::Installation.

In order to fix the crash, this patch makes the installationChanged()
callback cast user data to correct type.