Improve UI of proceed sheet
Right now if the text is really long, it gets elided at the bottom so
you need to click a button to see the rest, and the two distinct
sections of the text are kind of unclear.

This commit puts the text in the sheet's own scrollview so you can scroll
immediately, and adds a line break between the sections. It also puts the
bottom buttons in a footer, which looks better.

BUG: 419341
FIXED-IN: 5.20