Make all OverlaySheets window-modal
Right now half of Discover's OverlaySheets are view-modal and the other
half are window-modal. This feels arbitrary to the user, and view-modal
sheets look weird since the global drawer doesn't darken. This is
because it's still interactive, but there's really no point in being
able to interact with the global drawer while a sheet is open anyway. So
let's just make them all window-modal.

BUG: 426065
FIXED-IN: 5.20