Use a ListView for the screenshots gallery
This has the advantage of simplifying the code considerably. This also
unify the style with apps.kde.org and make the screenshot use the full
height of the containers, so they don't loook tiny anymore

We are losing the shadow affect on the screnshot, but considering that
the screenshots could have already shadow, this remove the double
shadow effect.

This also adds some navigation button on the desktop because horizontal
scrolling is not always an intuitive navigation pattern on a desktop.