Forward installer update error (so user is informed of their options)
As discovered by Piotr, attempting to update some KNewStuff originated
packages in Discover results in the update not in fact being updated,
without anything happening in Discover apart from the update just
staying around in the UI. This patch forwards an error from
KNSCore::Engine which gives some further information as to what the user
can do to perform this update.

This is a half-solution to the problem, as the TODO added in here
suggests, but it seems to me better to be explicit about the problem to
our users, rather than just not doing anything, and informing them how
they can fix it themselves. Until we have a full query system for
Discover's transactions, this seems likely to be the best we can do (and
actually implementing that system will take some time to get right, and
we don't want to just ignore the problem until them).

BUG:430812