Increase maximum width of review dialog to a sane size
A maximum value was set over a year ago but seems to have been somehow
ignored until recent changes in Kirigami made it take effect--at which
point it was revealed to be an absurdly low maximum value (144px wide
with the standard 10pt Noto Sans font). The principle of a maximum width
makes sense to prevent the dialog from looking silly with very wide
windows, but the current maximum is much too small. Let's triple it.

BUG: 432807
FIXED-in: 5.20.1