Mention backend name in "Install" button to make the source super clear
This allows the backend to be really really clear to the user, even if
they're not familiar with the logos and don't think to hover the cursor
over it to see the tooltip. In fact with this commit, the tooltip is
removed because the name is now visible by default in the button name,
so it's no longer needed.

BUG: 433370
FIXED-IN: 5.22