Make updates being available more obvious in compact mode
Prior to this patch, once Discover was started up and in compact mode,
there would be no visible indication at all that updates were available,
unless the user manually pulled out the global drawer and spotted the
orange entry at the bottom. This introduces an inline message at the
top, giving the user a positive bit of information that there are in
fact updates available, and what they can do about that.