flatpak: Specify when a source is user-local
A user might have added a repository twice both on the system and the
user deployments. This means that they'll be listed two seemingly equal
repositories and feel the world is melting.
This adds (user) at the end when it's a user-specific installation
(which generally are less common) so that they can be differentiated.

BUG: 436143