Clarify "Updates" Jump List Action
This action suffers from unclear text because it does not start with an
action verb, and als its icon is identical to that of the new "Software
Update" KCM, causing the two to look nearly identical in search results.
Also these kinds of actions are supposed to be using monochrome icons
anyway.

This commit fixes those issues.

BUG: 436141
FIXED-IN: 5.22