Only show "Successful Offline Update" notification on mobile
In general, our HIG recommends against showing notifications for success,
because the user expects actions that they initiate to succeed. However
apparently things are different in mobile-land, where user expectations
are different.

Therefore, we should only show the "Successful Offline Update"
notification for mobile, where people expect it, and not on the desktop,
where people do not.

BUG: 437681
FIXED-IN: 5.23