Move app delegate's source icon into the action button
Right now ApplicationDelegate shows an icon for the source for apps from
a non-default source. However the icon appears floating in the corner of
the delegate rather than inside the button to install or remove the app.
This was done as a workaround for https://bugs.kde.org/show_bug.cgi?id=433433
but feels kind of awkward.

That issue is not yet resolved, but there is an alternative: write a
custom content item for the button to add the missing feature. This lets
us put the icon in the button as originally intended without beeding to
wait for a fix in the framework.

This commit implements that change.