Be less aggressive about reboot notifications with offline updates
When we're doing online updates, it's imperative that the user reboot
ASAP, as the system is in an unknown state until that happens. However
with offline updates, this is not the case, and the system can safely be
used  indefinitely between the time when the update is downloaded and
the time when it's installed during the next reboot.

Therefore, when using offline updates, we don't need to badger the user
as much to reboot, so this commit disables the persistent "reboot now"
notification and adjusts the message shown on the updates page.
Hopefully this should alleviate some of the frustration that users have
expressed when using distros that enable offline updates by default.

BUG: 436929
BUG: 438443
BUG: 437454
FIXED-IN: 5.23