Move home button to the left of the sidebar
There's this pattern that is commonly used among most kirigami-based
applications: [searchField, hamburger menu] or [searchField, button]
but there's one (maybe more?) exception to this rule, System Settings,
which uses both a hamburger menu and a button in it's sidebar header
alongside the searchField.

In a recent change done to System Settings the hamburger menu and the
home button position were swapped ([menu, searchField, button] to
[button, searchField, button]) as most kirigami-based applications
tend to put the hamburger menu to the right on their sidebars.

This commit moves the home button to the left so it's consistent with
this pattern.

See: https://invent.kde.org/plasma/systemsettings/-/merge_requests/82