Make it easier to file distro bugs
Discover currently tells users to file bug reports against their distro
in a few places, but doesn't tell users how to do so, and in any event
is often is bit cagey about what exactly the error is.

As a result, we still often get bug reports against Discover that are
distro issues which we can't do anything about. This wastes developer
and bug triager time in KDE, frustrates users by giving them the
run-around, and makes Discover look bad in cases where the problem is
someone else's fault.

This commit attempts to improve that situation by making the error
messages clearer and also adding a big "Report this issue" button that
users can click on which will take them to the bug report URL for their
distro, not for Discover itself.

BUG: 442785
FIXED-IN: 5.24