Sources: Fix displaying sources
This is a work-around to 2 half-bugs:
- ListView doesn't really care for models with columns
- QConcatenateTablesProxyModel enforces that we see the minimum amount
of columns available in all of the models.

Empty models (like QStandardItemModel) will report 0 columns if they're
empty. This makes the resulting model to appear bogus to the view.
ListView doesn't seem to refresh the items when a new column is added
because it doesn't care about columns.

Result: this workaround after a bunch of wasted hours. :(

BUG: 442604