ApplicationScreenshots: Improve visuals and behavior
- Using new WheelHandler features instead of ScrollView to stop eating
wheel events when trying to scroll vertically.
- Depends on
https://invent.kde.org/frameworks/kirigami/-/merge_requests/415
- When scrolling vertically through the horizontal list view with a
touchpad, scrolling is really slow and seems to only move in small
increments. I have no idea why this is, but it has nothing to do with
ScrollView or WheelHandler. Even If you use neither, the issue remains,
so maybe it's a Flickable problem? Oddly, a mouse is perfectly fine.
- No more horizontal scrollbar. Having it there looked kind of ugly and
it didn't seem that necessary. The next/previous buttons indicate if
you're at the beginning or end. You can scroll horizontally with Alt + a
vertical wheel (like Qt Widgets), a horizontal wheel, a touchpad or by
flicking if you're not using a mouse/touchpad.
- Horizontal scrolling with a mouse should be less jittery now.
Touchpads are a bit bugged on Wayland though.
- Next/previous buttons in the ListView now jump to the image nearest to
or just outside of the edges.
- ListView now has some margins on the left and right in the parent
layout.
- Animations from clicking the left/right buttons are a bit faster and
rely on standard durations now.
- ListView is now pixelAligned
- Some accessibility info has been set. Not super descriptive since a
screenshots listview is mostly useless for blind people, it just says
what kinds of UI components are being used.
- Shortcuts are mirrored with mirrored layouts
- Using asynchronous for all images to improve page load times
- Smoother, faster and more RAM efficient screenshot thumbnails
- Navigation for popup and thumbnail list view are now separate
- You can now click anywhere to close the popup. The close button was
removed because it's kind of ugly. Maybe add a custom close button that
looks nice in a fullscreen overlay? Would be nice to use custom
next/previous buttons in the overlay too.
- No more background for the popup, instead using a darker overlay.
- Next/previous buttons in the listview disappear when the popup appears
so that they can't be confused with the identical looking popup
next/previous buttons.