Don't auto-accept in the search field
Search operations are network-intensive, so we can't have
search-as-you-type. Currently Discover is trying to do it anyway, by
auto-accepting after a 2.5 second delay. However this just isn't a good
user experience, and makes the accept interval seem random and it can
surprise you if you're not expecting it. When network limitations
dictate that results can't be updated in real time, we should just fall
back to making the user hit the return key.

BUG: 445142
FIXED-IN: 5.24