Improve user-friendliness of "Update Issue" message some more
This message still shoves technical details in the user's face, which
can be a scary wall of text that will prevent them from reading it or
anything beneath it--including the "Please report this issue to the
packagers of your distribution" text.

This commit turns it into a two-stage message: by default the user will
see a friendly error message that advises them to simply try again later
(which often does fix transient packaging issues), and if they want,
they can show the technical details and report it to their distro. When
they do so, the technical error message is shown in a TextArea that you
can easily copy text from, and only then is the "Report this issue"
button shown.

This should improve the UX for the "your distro is shipping broken
packages" case.