Flatpak: Fix loading apps from alien flatpakref files
So far it wasn't working well, we were relying on adding the source then
loading the resoruce which is a rather weird way to go for this.
Instead, now we are loading the appstream metadata locally to extract
the component and get all the metadata which gets eventually cleaned up.

BUG: 445596