PackageKitUpdater: Do not emit error when user cancels an update
A succesful cancelation is notified by the backend via the error
code `PackageKit::Transaction::ErrorTransactionCancelled`.

Do not emit an error message in this case, as it is the expected
outcome.

The same condition is used in PKTransaction.cpp:299

BEFORE:
![Screenshot_20220118_210009](/uploads/da7aeedcf88f67c3541fb3b61178951e/Screenshot_20220118_210009.png)


(cherry picked from commit 455087d469ccb92fd1589886c07bc40fd10bf639)