NavigationTabBar: Avoid loop on actions in tab group
The actions in Kirigami.NavigationTabBar are bounded by an
exclusive button group, which on start-up competes with the
binding on the `checked` property within those actions.

This results in an endless loop of alternating states,
and eventually a crash when the QV4 heap saturates.

Setting this group as non-exclusive fixes the issue.

BUG: 448460
FIXED-IN: 5.24
(cherry picked from commit 8b45c2042ef2518e28c6e98f0472af578d955ce7)