Show permissions a Flatpak has on application page
Flatpaks run inside an isolated and limiting sandbox. They usually
require additional resources on the system to work. Permissions to
access these resources outside the sandbox are granted to the
application by default, during installation. Discover does not list
which these permissions are, presently.

This commit lists these permissions out on the application's page. Each
permission is shown in a Kirigami.BasicListItem, which are laid out as a
vertical list below the Reviews' section by an instantiator. The model
to this instantiator resides in the new FlatpakPermission source/header
files along with the respective class.

Permissions are initially loaded from the metadata file of the
application, and then passed to the model constructor.