Warn users about the risks of proprietary software
Users may not be aware of the differences between proprietary and FOSS
apps, or even what "proprietary" means. Discover can help do a bit of
constructive user education here.

For proprietary apps, now we display the license name in orange text
and display a "What's the risk?" link that when clicked will give users
a little mini explanation about the dangers of proprietary software,
including a link to a relevant gnu.org page on the subject.