Fix searchfield focus on touch launch
We don't want to focus the search field by default when Discover is
launched with a touch, because then the virtual keyboard appears in the
user's face and covers up the content.

There is existing code to do this, but it's not quite working. We want
the magic condition to be toggling the `focus:` property, which is what
controls whether it has focus by default. We don't need to care about
this if the user takes an action to explicitly focus it; that's all
right.

(cherry picked from commit ff07b70a78469e4ef94d0c1a3b62ad8a16c32726)