Use a button instead of a link to open the "What's the risk?" sheet
Right now this sheet is opened by clicking on a link with the text
"What's the risk?"

We got some user feedback that links scare them because it makes them
think a web browser window will open and they don't want that, and also
that links in apps that don't open webpages are weird in general.

This commit changes the UI to use a button instead, which would be a
more typical way for users to make something happen within an app.

BUG: 451678
FIXED-IN: 5.24