pk: Consider multiple package ids for one upgradeable resource
One package may be installed with multiple architectures, and thus
the package name shall not be considered unique for the update
transaction.

There are a few ways to achieve that, but this is the one that
requires the less refactoring of the packagekit backend.

BUG: 444600


(cherry picked from commit 4907ad07eeaa100d425cd9a4fbcc64f1e30d6adc)