Show an Arch-specific message if no app backends are available
Arch developers have repeatedly expressed a preference for Arch users to
use the command line to manage software, rather than Discover. For this
reason, they deliberately package Discover in such a way that you will
have to go out of your way to figure out how to install an app backend.

Unfortunately, this means that Discover looks broken if you install it
with its default set of dependencies, because it won't be able to find
any apps. Novice Arch users don't understand that this is a deliberate
packaging choice in Arch, and instead blame Discover.

To alleviate this, we have introduced an error page that has become
ever mode specific over time about what exactly the problem is. Despite
this, novice Arch users remain confused about it. And the text we show
today is not really correct, as it implies that the issue is caused by
a bug in Arch's packaging. But it is *not* a bug; Arch packagers
package Discover this way deliberately, because they don't want Arch
users to use Discover to get apps. They consider this to be an expert
level feature that you shouldn't use unless you know what you're doing,
which inherently conflicts with the UX that we want to offer in
Discover.

Accordingly, for Arch users, let's show an Arch-specific message that
explains what's going on in clear and precise language, and offers an
actionable path forward.