ApplicationResourceButton: switch to individual ToolTip instances
We currently use a shared tooltip ostensibly to save memory, but the
shared opening/closing conditions cause the tooltips to close too early
when the cursor has hovered over more then one button by the time it
comes to rest on a different one.

To fix around this in the simplest way, let's switch to give each button
a unique ToolTip instance. In testing, there is no appreciable increase
in memory usage.

BUG: 456129
FIXED-IN: 5.25.3