Use correct variable to denote that ResourcesModel is done initializing
Due to somewhat confusing variable naming, we were accidentally setting
m_initializingBackends to false, when that variable actually holds the
number of backends that are still initializing. Instead we need to set
m_isInitializing to false.

This fixes a bug where the placeholder message for the MIME type screen
would fail to appear as expected, because one of the conditions in its
opacity property was `!ResourcesModel.isInitializing` and that was never
getting set to false.

BUG: 457528
FIXED-IN: 5.24.7