Flatpak: Never show both version number and branch name in one UI string
It can be terribly confusing since many apps and runtimes use branch
names that are in the form of a number. So to the user, it looks like
the app has two version numbers which doesn't make sense.

Instead, always show the version number if it's available. If not, only
show the branch name and hopfully it's in the form of a version number
so it looks like a version number to the user.

BUG: 458273
FIXED-IN: 5.26