Fwupd: don't show version number in description for only one update
If a fwupd update includes multiple releases because it's been a while
since you updated, then it makes sense to show the version number of
each invidividual one above that release's description.

But when there is only one release, it's redundant since the same
version number is already shown on the page's metadata bar on the
Application page, and the subtitle on the Updates page. So in this case
let's not show it.