Fix contentRatingDescription never being empty
Various parts of the QML check to see if the string returned by this
function is empty, but it never is because as long as you have a new
enough AppStream version, the returned value will always at least have
an asterisk and a newline.

This commit fixes that problem by generating the final string in the
loop, rather than at the end in the return statement. This allows for it
to remain empty if nothing happens in the loop.