Detect metered connection
On Qt 6.3, if the system supports it, check if the connection is metered instead of checking if it is Ethernet or Wi-Fi.
Both Ethernet and Wi-Fi can be metered, and if the computer is connected via a network bridge or some other more complex setup, it will return "unknown"
and automatic updates wouldn't happen.