Reviews: Don't check for name field being empty when it's not visible
Now, under certain circumstances, the name field will not be shown. But
the logic for ensuring all review data has been filled out was not
taking that into account, so it become impossible to submit a review
when the field was invisible because technically it was empty, so it
triggered the "Insert a name" condition.

This commit fixes that by disabling the name check when the name field
is invisible.

BUG: 460504
FIXED-IN: 5.26.1