Remove category labels from Application delegates
The App delegate has code to hide the category label when it's equal to
the page title, which handles the case of an app that shows
"Accessibility" as its category while the user has browsed to the
"Accessibility" category. However it broke because it was doing a total
match rather than a substring match, and while fixing it I found more
conceptual issues with displaying the categories at all, such as
inappropriately re-appearing when in a sub-category and when the app
has multiple categories.

This made me realize that the category labels on the delegate are little
more than visual noise that does not really help the user make a
decision or install an app. Because they're not adding value, let's
remove them. This also lets us reduce the height of non-compact
delegates a little bit to save space and increase information density.