Prefer openining the installed version of an app
When opening discover with a specific app as a command line argument and
this application has several versions, prefer showing the one that is
installed.

Since the app state might be delayed, we need to monitor the state of
the packages that don't know it for a bit. Delaying the initial state is
something we do commonly on the PackageKit backend.

BUG: 461664


(cherry picked from commit 706ebd9a87bbf9cdef090df97e28d38ea501bb54)