Increase startup parallelisation
On my system, it takes about 300ms to have a fully functional Discover
window. We were only starting to load the backends at around the 200ms,
this means that there was little chance to issue requests to the
(packagekit, separate threads, etc).

What we are doing here is to start the backends earlier (without
delaying registerAllBackends) so there is this opportunity indeed.

After this change I don't see a big difference, but it should help us
minimise the time where we show a progress bar as we are generally
waiting for these requests to get back to us.