Fix loading backends
Since now the loading of all backends happens as a block, the
ResourcesModel constructor was a big beast. This would make tons of
things happen before ResourcesModel::s_self was ever set.
This changes it so the constructor only sets the minimum necessary
things and the rest is done in a next call that will now happen just
once again.

BUG: 462447