ApplicationResouceButton: place icon side-by-side to the title
Currently we use a ColumnLayout to place the icon, title and subtitle.
This is visually awkward since it left aligns the relatively small icon
where all the text follows us a blob.

Instead use a mix of Column and RowLayout, so the icon and title are
side-by-side, while the subtitle spans across the full width.

Spacing and similar numbers were gracously copied from plasma-welcome's
PlasmaFeatureButton widget.

With that change, we need to tweak the stacked layout calculation to
cater for the icon side + extra spacing.

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>