Fix case where an app provides an appstream id
When an app provides an appstream id, when the user clicks to select it, we will make an appstream search specifying both the app id and any ids that it provides
setResources usually disallows duplicates, and that helps catch bugs. However, in this case, duplicates are expected since we are searching for two ids of the same component (i.e. org.gimp.GIMP and gimp.desktop)