Fix plural strings in search result header text
We don't always have more than one result, and in some languages, the
singular form is used for numbers like 21. So use i18np() here.

The orders of the arguments is reversed in some function calls because
i18np() requires that that the number which determines a string's
plurality be the first argument.