Help users stuck in empty category searches
If the user is browsing a category, Discover any limits search results
to only things that match within that category. But if the user does
not realize they are in a category, and they search for something they
know to exist, but it lives in another category, Discover will
currently find nothing and look stupid, and the user will be frustrated.

To alleviate this, let's change the message shown when nothing is found
for a category search. Instead of saying, "we couldn't find it in the
available sources" (which is false anyway), Discover will now say that
it simply couldn't find it in the current category, and will show a
button to let the user search globally.

BUG: 462809
FIXED-IN: 5.27