flatpak: Make sure we are reading the correct metadata
Also fetch the metadata for the apps that are available on the repo but
have different branches to choose from.

Since the branches model is not something appstream accounts for, so we
all get confused. Therefore, we don't look for it in the metainfo file from
the app and just add the bundle to it so it fits with the rest of components.

BUG: 464128