ui: Fix plural calculation of app list page heading
We cannot use roughCount as the value for i18np(). It's a string and we
need a number.

As it's <1000 JS will be overly smart and convert the number into a string
but eventually we give "1000+" and then the conversion returns 0. Without
warnings. 👍

BUG: 464882


(cherry picked from commit 9462a8164369944e1e1f8d46f07c861c77c9e73d)