flatpak: Fix warning of dupd identifiers in the url
We would get the following URL for VLC:
appstream://org.videolan.vlc?alt=org.videolan.vlc,vlc.desktop

This is because the appstream id is org.videolan.VLC but QUrl always
makes the host lowercase.
This changes to put the id in the path part of the URL so we can treat
them separately in AppStreamUtils::appstreamIds, which is the
centralised place we have to extract ids in Discover. Now the URL we
generate will be:
appstream:org.videolan.VLC?alt=org.videolan.vlc,vlc.desktop