screenshots: Only use AnimatedImage if we think there's a chance
Not long ago we switched to using AnimatedImage to show every
screenshot.
It kind of works but not entirely, especially on sequential streams.
Only use the AnimatedImage in cases we know it's supposed to be a video.
This can be specified in AppStream's metadata.

While this could be seen as an otherwise error-prone change, it makes
the
API much more solid by not juggling two lists taht need to be the same
size.