ApplicationPage: Improve narrow layout with lots of buttons
When there are a lot of ApplicationResourceButton instances in the
layout, they can become very difficult to read when the window is in
mobile mode because they become tall and skinny and the text wraps
weirdly.

To fix this, switch to a stacked layout whenever any button's title text
would otherwise wrap onto the next row.