Always show distro name for PackageKit apps
Discover tries to show the name of the repo that an app comes from. This
string is generally human-readable for Snap, Flatpak, KNS, and
even apps from 3rd-party PackageKit repos. But it is usually not
human-readable for apps that come from distro repos, because it's common
for distros to ship multiple repos that serve different purposes. When
we try to show the name of these repos in the UI, it's nonsense; we end
up with text like "Distributed by: xenial-security universe" and
Distributed by: updates". This text doesn't make any sense to the user
and can even be scary; imagine "Distributed by: Restricted" which is
what you can see on Ubuntu-based distros!

The user does not care about which specific distro repo the app comes
from; what they care about is that it came from their distro.

What would be ideal is if PackageKit could tell us whether a particular
PackageKit repo was 1st-party or 3rd-party. Then for 3rd-party repos we
could show the actual name (which usually ends up looking reasonably
human-readable, like "Packman" or "rpmfusion-free"), and for 1st-party
repos, we could instead substitute the distro name.

Unfortunately PackageKit doesn't have that functionality right now. See
https://github.com/PackageKit/PackageKit/issues/607.

So for now, just show the distro name for all PackageKit apps. This
fixes the nonsense strings issues, and while it does result in
inaccuracy for users who have installed 3rd-party package repos, those
users are generally technically knowledgeable enough to understand.

BUG: 465204
FIXED-IN: 5.27.2